+++ angular/projects/catalyst/src/lib/tooltip/tooltip.directive.ts
@@ -1,4 +1,4 @@
-import { Directive, Input, OnInit, TemplateRef, ViewContainerRef } from '@angular/core';
+import { ComponentRef, Directive, Input, OnChanges, OnInit, TemplateRef, ViewContainerRef } from '@angular/core';
 import { CatTooltip } from '../directives/proxies';
 
 /**
@@ -8,8 +8,10 @@ import { CatTooltip } from '../directives/proxies';
   selector: '[catTooltip]',
   standalone: false
 })
-export class CatTooltipDirective implements OnInit {
+export class CatTooltipDirective implements OnInit, OnChanges {
   @Input() catTooltip!: string;
+
+  private component?: ComponentRef<CatTooltip>;
 
   constructor(
     private templateRef: TemplateRef<unknown>,
@@ -17,8 +19,15 @@ export class CatTooltipDirective implements OnInit {
   ) {}
 
   ngOnInit() {
-    this.viewContainer.createComponent(CatTooltip, {
+    this.component = this.viewContainer.createComponent(CatTooltip, {
       projectableNodes: [this.viewContainer.createEmbeddedView(this.templateRef).rootNodes]
-    }).instance.content = this.catTooltip;
+    });
+    this.component.instance.content = this.catTooltip;
+  }
+
+  ngOnChanges() {
+    if (this.component) {
+      this.component.instance.content = this.catTooltip;
+    }
   }
 }