+++ core/stencil.config.ts
@@ -54,7 +54,8 @@ export const config: Config = {
   sourceMap: true,
   globalStyle: 'src/styles/index.scss',
   extras: {
-    enableImportInjection: true
+    enableImportInjection: true,
+    tagNameTransform: true
   },
   outputTargets: [
     {